﻿TMXParser can now parse polygon and polylines pixel points. Access said points via TMXObject.
The TMXObject can now have polygon and polyline points which can be accessed via relative methods.
The TMXParser can detect the lines and passes to the last TMXObject to parse.
SAXUtilsObject exists to parse the pixel points into a 2D int array.
TMXConstants updated to hold the required XML tags.